docs and bug fixing
