+++ config.cfg
@@ -9,6 +9,9 @@ secret_key=YouMustChangeThisValueToSomething.
 token_duration=3600
 cors_origin=*
 
+[Git]
+
+
 [AWS]
 bucket=public-eo-data
 access_key_id=SET THIS TO YOUR ACTUAL VALUE USING ENV VARS
